Stop favicons being hashed (#60)
Favicons are not showing up in the google search results for docs and this material site. It's possible this is because the favicons are being converted into a file name with a hash e.g. favicon_233242.ico. This ensure they are in the public folder where Astro won't hash them and change the name. This might fix the google search results issue.